pyproject.toml
[tool.poetry]
name = "webapp"
package-mode = false 
version = "0.1.0"
description = ""
authors = ["Charlie00CR"]

[tool.poetry.dependencies]
sqlalchemy = "^1.4.0"
pymongo = "^4.4.1"
python = "^3.11"
uvicorn = "^0.30.5"
fastapi = "^0.103.0"  
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
python-keycloak = "^4.3.0"
requests = "2.31.0"
pydantic = "2.8.2"
python-multipart = "^0.0.5"
redis = {extras = ["hiredis"], version = "^5.1.1"}

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [
    "."
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

